+++ src/onegov.event/onegov/event/models/__init__.py
@@ -1,8 +1,9 @@
-from onegov.event.models.event import Event
+from onegov.event.models.event import Event, EventFile
 from onegov.event.models.occurrence import Occurrence
 
 
 __all__ = (
     'Event',
+    'EventFile',
     'Occurrence',
 )